pcengines/apu1: Remove unused smbus.asl
The smbus.asl operation regions prevent the Linux i2c driver (i2c_piix4)
for this chipset from claiming the ioport ranges and thus it fails to
load.

The methods defined in smbus.asl are not used in the DSDT and also don't
exist in the DSDT of the vendor firmware.

In particular due to the following check in i2c-piix4.c will fail unless
acpi_enforce_resources=no is explicitly set on the Linux kernel
parameters:
  if (acpi_check_region(piix4_smba, SMBIOSIZE, piix4_driver.name))
          return -ENODEV;

Depending on kernel options the only error message printed is

  ACPI Warning: SystemIO range 0x0000000000000B00-0x0000000000000B07 conflicts with OpRegion 0x0000000000000B00-0x0000000000000B0B (\SMB0) (20150410/utaddress-254)
  ACPI: If an ACPI driver is available for this device, you should use it instead of the native driver

However since it does not implement a standard interface there is no
native ACPI driver for smbus.asl.

Change-Id: Id8401e8b36f0e2412d490a92c20540a04d853125
Signed-off-by: Tobias Diedrich <ranma+coreboot@tdiedrich.de>
Reviewed-on: http://review.coreboot.org/10539
Tested-by: build bot (Jenkins)
Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>